Move GraphQL stuff to its own module
